refactored code and added documentation
