+++ app/src/main/java/com/jrod7938/textchangeapp/navigation/AppScreens.kt
@@ -1,5 +1,41 @@
+/*
+ * BSD 3-Clause License
+ *
+ * Copyright (c) 2023, Jrod7938, Khang-ALe, jesma14, Holesum
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of the copyright holder nor the names of its
+ *    contributors may be used to endorse or promote products derived from
+ *    this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
 package com.jrod7938.textchangeapp.navigation
 
+/**
+ * Enum class to represent the different screens of the app
+ *
+ * @constructor Creates an enum class to represent the different screens of the app
+ */
 enum class AppScreens {
     SplashScreen,
     LoginScreen,
@@ -10,8 +46,15 @@ enum class AppScreens {
     SellBookScreen,
     SavedBooksScreen;
 
-    companion object{
-        fun fromRoute(route: String): AppScreens = when(route?.substringBefore("/")) {
+    companion object {
+        /**
+         * Get the screen from the route
+         *
+         * @param route route to get the screen from
+         *
+         * @return AppScreens
+         */
+        fun fromRoute(route: String): AppScreens = when (route?.substringBefore("/")) {
             SplashScreen.name -> SplashScreen
             LoginScreen.name -> LoginScreen
             CreateAccountScreen.name -> CreateAccountScreen
