+++ app/src/main/java/com/jrod7938/textchangeapp/screens/login/LoginScreen.kt
@@ -31,59 +31,29 @@
 
 package com.jrod7938.textchangeapp.screens.login
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material3.Button
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
-import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.platform.LocalSoftwareKeyboardController
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
-import com.jrod7938.textchangeapp.R
 import com.jrod7938.textchangeapp.components.AppLogo
-import com.jrod7938.textchangeapp.components.EmailInput
+import com.jrod7938.textchangeapp.components.UserForm
 import com.jrod7938.textchangeapp.navigation.AppScreens
 
 
@@ -149,135 +119,5 @@ fun LoginScreen(
                 color = MaterialTheme.colorScheme.secondary
             )
         }
-    }
-}
-
-@OptIn(ExperimentalComposeUiApi::class)
-@Preview(showBackground = true)
-@Composable
-fun UserForm(
-    loading: Boolean = false,
-    isCreateAccount: Boolean = false,
-    onDone: (String, String) -> Unit = { email, pwd -> }
-){
-    val email = rememberSaveable { mutableStateOf("") }
-    val password = rememberSaveable { mutableStateOf("") }
-    val passwordVisibility = rememberSaveable { mutableStateOf(false) }
-    val passwordFocusRequest = remember { FocusRequester() }
-    val keyboardController = LocalSoftwareKeyboardController.current
-    val valid = remember(email.value, password.value) {
-        email.value.trim().isNotEmpty()
-                && email.value.contains("@pride.hofstra.edu")
-                && password.value.trim().isNotEmpty()
-                && password.value.length >= 6
-    }
-
-    val modifier = Modifier
-        .height(250.dp)
-        .background(color = MaterialTheme.colorScheme.background)
-        .verticalScroll(rememberScrollState())
-
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        if (isCreateAccount) {
-            Text(
-                text = stringResource(id = R.string.create_acct),
-                modifier = Modifier.padding(4.dp)
-            )
-        } else {
-            Text(
-                text = "Welcome, please login to continue!",
-                modifier = Modifier.padding(4.dp)
-            )
-        }
-        EmailInput(
-            emailState = email,
-            enabled = !loading,
-            onAction = KeyboardActions { passwordFocusRequest.requestFocus() }
-        )
-        PasswordInput(
-            modifier = Modifier.focusRequester(passwordFocusRequest),
-            passwordState = password,
-            labelId = "Password",
-            enabled = !loading,
-            passwordVisibility = passwordVisibility,
-            onAction = KeyboardActions {
-                if(!valid) return@KeyboardActions
-                onDone(email.value.trim(), password.value.trim())
-            }
-        )
-        SubmitButton(
-            textId = if(isCreateAccount) "Create Account" else "Login",
-            loading = loading,
-            validInputs = valid,
-        ){
-            onDone(email.value.trim(), password.value.trim())
-            keyboardController?.hide()
-        }
-
-    }
-}
-
-@Composable
-fun SubmitButton(
-    textId: String,
-    loading: Boolean,
-    validInputs: Boolean,
-    onClick: () -> Unit
-) {
-    Button(
-        modifier = Modifier
-            .padding(10.dp)
-            .fillMaxWidth(),
-        enabled = !loading && validInputs,
-        shape = CircleShape,
-        onClick = onClick
-    ) {
-        if (loading) CircularProgressIndicator(modifier = Modifier.size(25.dp))
-        else Text(text = textId, modifier = Modifier.padding(5.dp))
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun PasswordInput(
-    modifier: Modifier,
-    passwordState: MutableState<String>,
-    labelId: String,
-    enabled: Boolean,
-    passwordVisibility: MutableState<Boolean>,
-    imeAction: ImeAction = ImeAction.Done,
-    onAction: KeyboardActions = KeyboardActions.Default
-) {
-    val visualTransformation = if(passwordVisibility.value) VisualTransformation.None else PasswordVisualTransformation()
-    OutlinedTextField(
-        modifier = modifier
-            .padding(10.dp)
-            .fillMaxWidth(),
-        value = passwordState.value,
-        label = { Text(text = labelId) },
-        singleLine = true,
-        onValueChange = { passwordState.value = it },
-        textStyle = TextStyle(
-            fontSize = 18.sp,
-            color = MaterialTheme.colorScheme.onBackground),
-        enabled = enabled,
-        keyboardOptions = KeyboardOptions(
-            keyboardType = KeyboardType.Password,
-            imeAction = imeAction),
-        visualTransformation = visualTransformation,
-        trailingIcon = { PasswordVisibility(passwordVisibility = passwordVisibility)},
-        keyboardActions = onAction
-    )
-}
-
-@Composable
-fun PasswordVisibility(passwordVisibility: MutableState<Boolean>) {
-    val visible = passwordVisibility.value
-    IconButton(onClick = { passwordVisibility.value = !visible}) {
-        Icons.Default.Close
     }
 }
